fix not deleting files on the backend; finish UI

+++ frontend/src/app/core/components/explorer/state/explorer.state.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Navigate } from '@ngxs/router-plugin';
 import { Action, Selector, State, StateContext } from '@ngxs/store';
-import { append, patch, removeItem } from '@ngxs/store/operators';
+import { append, patch, removeItem, updateItem } from '@ngxs/store/operators';
 import { DirectoriesService, FilesService } from 'generated';
 import { EMPTY, expand, map, of, switchMap, takeLast, tap } from 'rxjs';
 import { ExplorerActions } from 'src/app/core/components/explorer/state/explorer.actions';
@@ -93,8 +93,23 @@ export class ExplorerState {
 	}
 
 	@Selector()
+	public static getContent(state: ExplorerStateModel) {
+		return state.tree[state.directory] ?? [];
+	}
+
+	@Selector()
 	public static getTree(state: ExplorerStateModel) {
 		return state.tree;
+	}
+
+	@Action(ExplorerActions.SetDirectory)
+	public setDirectory(ctx: StateContext<ExplorerStateModel>, action: ExplorerActions.SetDirectory) {
+		ctx.setState(
+			patch({
+				directory: action.id,
+				isRoot: ROOT_ID === action.id,
+			})
+		);
 	}
 
 	@Action(ExplorerActions.Open)
@@ -125,21 +140,14 @@ export class ExplorerState {
 			return;
 		}
 
-		ctx.setState(
-			patch({
-				directory: action.id,
-				isRoot: action.id === ROOT_ID,
-			})
-		);
-
 		return ctx.dispatch([new ContentListActions.UnselectAll(), new Navigate(['explorer', action.id])]);
 	}
 
 	@Action(ExplorerActions.LoadInitialContent)
-	public loadInitialContent(ctx: StateContext<ExplorerStateModel>, action: ExplorerActions.LoadInitialContent) {
+	public loadInitialContent(ctx: StateContext<ExplorerStateModel>) {
 		const mergeTrees = (a: Tree, b: Tree) => ({ ...a, ...b });
 
-		return of({ tree: {}, crumbs: [], id: action.id }).pipe(
+		return of({ tree: {}, crumbs: [], id: ctx.getState().directory }).pipe(
 			expand((d) =>
 				!d.id
 					? EMPTY
@@ -170,6 +178,7 @@ export class ExplorerState {
 	@Action(ExplorerActions.LoadContent)
 	public loadContent(ctx: StateContext<ExplorerStateModel>, action: ExplorerActions.LoadContent) {
 		if (action.id in ctx.getState().tree) {
+			console.log('Skipped because already in state');
 			return;
 		}
 
@@ -195,12 +204,26 @@ export class ExplorerState {
 	@Action(ExplorerActions.CreateDirectory)
 	public createDirectory(ctx: StateContext<ExplorerStateModel>, action: ExplorerActions.CreateDirectory) {
 		const parentId = ctx.getState().directory;
+		const grandparentId = Object.keys(ctx.getState().tree).find((key) => ctx.getState().tree[key].some((item) => item.id === parentId));
 
 		return this.directoriesService.create(parentId, action).pipe(
 			switchMap((body) =>
 				this.directoriesService.getMetadata(body.id).pipe(
 					map((metadata) => ({ id: body.id, ...metadata })),
-					tap((metadata) => ctx.setState(patch({ tree: patch({ [parentId]: append([{ type: Type.Directory, ...metadata }]) }) })))
+					tap((metadata) =>
+						ctx.setState(
+							patch({
+								tree: patch(
+									grandparentId
+										? {
+												[grandparentId]: updateItem((item) => item.id === parentId, patch({ directories: (amt) => amt + 1 })),
+												[parentId]: append([{ type: Type.Directory, ...metadata }]),
+											}
+										: { [parentId]: append([{ type: Type.Directory, ...metadata }]) }
+								),
+							})
+						)
+					)
 				)
 			),
 			tap(() => ctx.dispatch(new ExplorerActions.HideCreateDirectoryComponent()))
@@ -209,13 +232,23 @@ export class ExplorerState {
 
 	@Action(ExplorerActions.DeleteDirectory)
 	public deleteDirectory(ctx: StateContext<ExplorerStateModel>, action: ExplorerActions.DeleteDirectory) {
+		const parentId = ctx.getState().directory;
+		const grandparentId = Object.keys(ctx.getState().tree).find((key) => ctx.getState().tree[key].some((item) => item.id === parentId));
+
 		return this.directoriesService._delete(action.id).pipe(
 			tap(() =>
 				ctx.setState(
 					patch({
-						tree: patch({
-							[ctx.getState().directory]: removeItem((item) => item.id === action.id),
-						}),
+						tree: patch(
+							grandparentId
+								? {
+										[grandparentId]: updateItem((item) => item.id === parentId, patch({ directories: (amt) => amt - 1 })),
+										[parentId]: removeItem((item) => item.id === action.id),
+									}
+								: {
+										[parentId]: removeItem((item) => item.id === action.id),
+									}
+						),
 					})
 				)
 			)
@@ -225,6 +258,7 @@ export class ExplorerState {
 	@Action(ExplorerActions.UploadFile)
 	public uploadFile(ctx: StateContext<ExplorerStateModel>, action: ExplorerActions.UploadFile) {
 		const directoryId = ctx.getState().directory;
+		const grandparentId = Object.keys(ctx.getState().tree).find((key) => ctx.getState().tree[key].some((item) => item.id === directoryId));
 
 		return this.filesService.upload(directoryId, action.file).pipe(
 			switchMap((body) =>
@@ -233,9 +267,17 @@ export class ExplorerState {
 					tap((metadata) =>
 						ctx.setState(
 							patch({
-								tree: patch({
-									[directoryId]: append([{ type: Type.File, ...metadata }]),
-								}),
+								tree: patch(
+									grandparentId
+										? {
+												[grandparentId]: updateItem(
+													(item) => item.id === directoryId,
+													patch({ files: (amt) => amt + 1, size: (s) => s + metadata.size })
+												),
+												[directoryId]: append([{ type: Type.File, ...metadata }]),
+											}
+										: { [directoryId]: append([{ type: Type.File, ...metadata }]) }
+								),
 							})
 						)
 					)
@@ -247,14 +289,31 @@ export class ExplorerState {
 	@Action(ExplorerActions.DeleteFile)
 	public deleteFile(ctx: StateContext<ExplorerStateModel>, action: ExplorerActions.DeleteFile) {
 		const directoryId = ctx.getState().directory;
+		const size = ctx.getState().tree[directoryId].find((item) => item.id === action.id)?.size;
+
+		if (!size) {
+			return;
+		}
+
+		const grandparentId = Object.keys(ctx.getState().tree).find((key) => ctx.getState().tree[key].some((item) => item.id === directoryId));
 
 		return this.filesService.deleteFile(directoryId, action.id).pipe(
 			tap(() =>
 				ctx.setState(
 					patch({
-						tree: patch({
-							[directoryId]: removeItem((item) => item.id === action.id),
-						}),
+						tree: patch(
+							grandparentId
+								? {
+										[grandparentId]: updateItem(
+											(item) => item.id === directoryId,
+											patch({ files: (amt) => amt - 1, size: (s) => s - size })
+										),
+										[directoryId]: removeItem((item) => item.id === action.id),
+									}
+								: {
+										[directoryId]: removeItem((item) => item.id === action.id),
+									}
+						),
 					})
 				)
 			)
